requirements.txt
astroid==3.3.8
coverage==7.6.12
dill==0.3.9
Faker==36.1.1
iniconfig==2.0.0
isort==6.0.1
mccabe==0.7.0
packaging==24.2
platformdirs==4.3.6
pluggy==1.5.0
pylint==3.3.4
pytest==8.3.5
pytest-cov==6.0.0
pytest-pylint==0.21.0
python-dotenv==1.0.1
tomlkit==0.13.2
tzdata==2025.1
